+++ web/src/middleware/mod.rs
@@ -297,6 +297,8 @@
 pub mod compress;
 #[cfg(any(feature = "compress-br", feature = "compress-gz", feature = "compress-de"))]
 pub mod decompress;
+#[cfg(feature = "logger")]
+pub mod logger;
 #[cfg(feature = "rate-limit")]
 pub mod rate_limit;
 #[cfg(feature = "tower-http-compat")]
@@ -308,7 +310,7 @@ pub mod limit;
 #[cfg(not(target_family = "wasm"))]
 pub mod sync;
 
-pub use xitca_http::util::middleware::{Extension, Logger};
+pub use xitca_http::util::middleware::Extension;
 pub use xitca_service::middleware::{Group, UncheckedReady};
 
 #[cfg(test)]